🚚 Move the TabStyling component
