+++ modules/ppcp-settings/resources/js/Components/Screens/Settings/Tabs/index.js
@@ -3,7 +3,7 @@ import { __ } from '@wordpress/i18n';
 import TabOverview from '../../Overview/TabOverview';
 import TabPaymentMethods from '../../Overview/TabPaymentMethods';
 import TabSettings from '../../Overview/TabSettings';
-import TabStyling from '../../Overview/TabStyling';
+import TabStyling from './TabStyling';
 import TabPayLaterMessaging from '../../Overview/TabPayLaterMessaging';
 
 /**
